Revert "build(deps): upgrade packages to minor and patch versions"
Error:
- Found errors in the terminal when run the `npm install`
npm ERR! Could not resolve dependency:
npm ERR! peer typescript@">=4.9.3 <5.2" from @angular-devkit/build-angular@16.2.5

Reason:
- The `ERESOLVE` error from npm is indicating that there is a conflict with the project dependencies. The main problem here seems to be a conflict between the version of `typescript` you have installed and what `@angular-devkit/build-angular` expects.
In short, `@angular-devkit/build-angular@16.2.5` expects a `typescript` version between `>=4.9.3` and `<5.2`, but you currently have `5.2.2` version of `typescript` installed.

Solution:
- Downgrade TypeScript to 5.1.6 instead use 5.2.2

Other possible solutions:
- I would recommend trying these solutions in the order they are presented, as the first option is the safest. The option to use `--legacy-peer-deps` should be your last resort, as it may lead to unexpected problems later on.
1. **Downgrade TypeScript**:
If you don't have a specific reason for using TypeScript 5.2.2, you can downgrade to resolve the conflict:
````bash
npm install typescript@5.1.6 --save-dev
```
This will install ``5.1.6` version of TypeScript, which is compatible with `@angular-devkit/build-angular@16.2.5`.
2. **Use Flag `--legacy-peer-deps`**:
If you do not want to change the version of TypeScript and are willing to risk incompatibilities, you can use the `--legacy-peer-deps` flag:
````bash
npm install --legacy-peer-deps
```
However, be careful when doing this. While it may resolve the error and allow you to install the dependencies, you could face runtime problems due to incompatibilities.
3. **Update `@angular-devkit/build-angular`** (if a newer version is available):
There may be a newer version of `@angular-devkit/build-angular` that is compatible with TypeScript 5.2.2. You can try upgrading:
````bash
npm install @angular-devkit/build-angular@latest --save-dev
```
Then, check if the problem is solved.
4. **Check your `peerDependencies`**:
If the project was cloned from a repository or is a previous project, it is possible that the `peerDependencies` in your `package.json` are causing the conflict. In that case, check and adjust the versions in the `package.json` to make sure they are compatible.

References:
- https://github.com/FusionAuth/fusionauth-quickstart-javascript-angular-web/issues/4
- This reverts commit 9e6f851e35e9f870fbf32a332c9189a215269de7.